Cad Aluno

+++ aula32.py
@@ -22,7 +22,7 @@ baseando-se no horário descrito, exiba a saudação apropriada. Ex.
 Bom dia 0-11, boa tarde 12-17, boa noite 18-23
 '''
 nome = input('Digite seu nome: ')
-hora = int(input('Digite a hora: '))
+hora = float(input('Digite a hora: '))
 if 0<= hora <=11:
     print(f'Bom dia {nome}!')
 elif 12<= hora <= 18:
